+++ config/routes.rb
@@ -21,6 +21,9 @@ Rails.application.routes.draw do
 
   namespace :admin do
     resources :replies, only: :index
+    get :spam_inbox, path: "spam-inbox"
+    post :unshadowban, path: "unshadowban/:id"
+    post :empty_spam, path: "empty-spam"
     resources :categories
     resources :exports do
       member do